backend/python/requirements.txt
asgiref==3.5.0
Django==4.0.1
djangorestframework==3.13.1
pytz==2021.3
PyYAML==6.0
sqlparse==0.4.2
uritemplate==4.1.1
